fix: Ensure that front and back are on separate pages
The generating script did not put \newpage between front and back page
because it seemed to be not necessary. However, then on the last page
front and back table may end up on the same page, e.g. when they have
only one row. This is fixed by putting \newpage between them.