pdfs: Replace static physical card page with generated ones
Now that the script is done, the clunky static LaTeX document can be
replaced with newer ones. The benefits are better sized cards by being
able to switch to almost any RxC layout and the use of the generated
SVGs for crisp resolution at any zoom.

While in principle the PDFs can be easily reproduced, they are added to
allow making use of them without running any code. After all, this
repository targets cubers, not coders.